Remove gtm when user removes preferences
