+++ src/shared/components/ConsentPreferences/index.tsx
@@ -2,16 +2,13 @@ import * as React from 'react';
 import { Button } from 'antd';
 import useLocalStorage from '../../hooks/useLocalStorage';
 
-const ConsentPreferences: React.FunctionComponent<{}> = () => {
-  const [consent, setConsent] = useLocalStorage<{
+const ConsentPreferences: React.FunctionComponent<{
+  consent?: {
     consentToTracking: boolean;
     hasSetPreferences: boolean;
-  }>('consentToTracking');
-
-  const onClickRemove = () => {
-    setConsent(undefined);
   };
-
+  onClickRemove?(): void;
+}> = ({ consent, onClickRemove }) => {
   return (
     <Button
       onClick={onClickRemove}
