+++ src/shared/layouts/MainLayout.tsx
@@ -5,12 +5,14 @@ import { push } from 'connected-react-router';
 import { NexusClient, Identity, Realm } from '@bbp/nexus-sdk';
 import { useNexus } from '@bbp/react-nexus';
 import { UserManager } from 'oidc-client';
+import { notification, Button } from 'antd';
 
 import Header, { ServiceVersions } from '../components/Header';
 import getUserManager from '../../client/userManager';
 import { getLogoutUrl, getDestinationParam } from '../utils';
 import { RootState } from '../store/reducers';
 import { version, url as githubIssueURL } from '../../../package.json';
+import useLocalStorage from '../hooks/useLocalStorage';
 
 import './MainLayout.less';
 import ConsentContainer from '../containers/ConsentContainer';
@@ -47,6 +49,63 @@ const MainLayout: React.FunctionComponent<MainLayoutProps> = ({
     localStorage.removeItem('nexus__state');
     userManager && userManager.signoutRedirect();
   };
+
+  const [consent, setConsent] = useLocalStorage<{
+    consentToTracking: boolean;
+    hasSetPreferences: boolean;
+  }>('consentToTracking');
+
+  const trackingConsentNotification = (
+    onClose: (accepted: boolean) => void
+  ) => {
+    const key = `tracking-consent-notification`;
+    const btn = (
+      <div>
+        <Button
+          size="small"
+          onClick={() => {
+            notification.close(key);
+            onClose(false);
+          }}
+        >
+          Don't Allow
+        </Button>{' '}
+        <Button
+          type="primary"
+          size="small"
+          onClick={() => {
+            notification.close(key);
+            onClose(true);
+          }}
+        >
+          Allow
+        </Button>
+      </div>
+    );
+
+    notification.open({
+      btn,
+      key,
+      onClose: () => {
+        // If the user dismisses the notification, assume they don't want tracking enabled.
+        onClose(false);
+      },
+      message: 'Send data & statistics to the developers?',
+      description: `
+        Send data to the developers in order to improve Nexus Web by tracking your activity.
+      `,
+      duration: null, // don't auto-close
+    });
+  };
+
+  if (!consent || !consent.hasSetPreferences) {
+    trackingConsentNotification(allow => {
+      setConsent({
+        consentToTracking: allow,
+        hasSetPreferences: true,
+      });
+    });
+  }
 
   // Remove version from API URL
   const splits = apiEndpoint.split('/');
@@ -101,8 +160,10 @@ const MainLayout: React.FunctionComponent<MainLayoutProps> = ({
         version={version}
         githubIssueURL={githubIssueURL}
         serviceVersions={versions.data}
+        consent={consent}
+        onClickRemoveConsent={() => setConsent(undefined)}
       />
-      <ConsentContainer trackingCode={gtmCode || ''} />
+      <ConsentContainer trackingCode={gtmCode || ''} consent={consent} />
       <div className="MainLayout_body">{children}</div>
     </>
   );
